requirements.txt
langchain>=0.3.27
langchain-community>=0.3.27
langchain-anthropic>=0.3.17
langchain-chroma>=0.2.5
langchain-google-genai>=2.1.8
langchain-google-vertexai>=2.0.27
langchain-openai>=0.3.28
langchain-text-splitters>=0.3.8
langchain-tavily>=0.2.10
langgraph>=0.5.4
langgraph-checkpoint>=2.1.1
langgraph-cli[inmem]>=0.3.6
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
tqdm>=4.65.0
chromadb>=1.0.9
python-dotenv>=1.0.0
requests>=2.31.0
langsmith>=0.3.45 